- completed terminal (loading libraries) - code cleanup

+++ src/main/com/mucommander/commons/file/impl/sevenzip/provider/SevenZip/LzmaBench.java
@@ -46,7 +46,7 @@ public class LzmaBench
 			numBits -= NumBits;
 			result = (Value << numBits);
 			Value = RG.GetRnd();
-			result |= Value & (((int)1 << numBits) - 1);
+			result |= Value & ((1 << numBits) - 1);
 			Value >>>= numBits;
 			NumBits = 32 - numBits;
 			return result;
@@ -73,7 +73,7 @@ public class LzmaBench
 		int GetLogRandBits(int numBits)
 		{
 			int len = RG.GetRnd(numBits);
-			return RG.GetRnd((int)len);
+			return RG.GetRnd(len);
 		}
 		int GetOffset()
 		{
@@ -81,8 +81,8 @@ public class LzmaBench
 				return GetLogRandBits(4);
 			return (GetLogRandBits(4) << 10) | RG.GetRnd(10);
 		}
-		int GetLen1() { return RG.GetRnd(1 + (int)RG.GetRnd(2)); }
-		int GetLen2() { return RG.GetRnd(2 + (int)RG.GetRnd(2)); }
+		int GetLen1() { return RG.GetRnd(1 + RG.GetRnd(2)); }
+		int GetLen2() { return RG.GetRnd(2 + RG.GetRnd(2)); }
 		public void Generate()
 		{
 			RG.Init();
@@ -237,7 +237,7 @@ public class LzmaBench
 	{
 		long t = GetLogSize(dictionarySize) - (18 << kSubBits);
 		long numCommandsForOne = 1060 + ((t * t * 10) >> (2 * kSubBits));
-		long numCommands = (long)(size) * numCommandsForOne;
+		long numCommands = (size) * numCommandsForOne;
 		return MyMultDiv64(numCommands, elapsedTime);
 	}
 	
@@ -370,7 +370,7 @@ public class LzmaBench
 				if (crcOutStream.GetDigest() != crc.GetDigest())
 					throw (new Exception("CRC Error"));
 			}
-			long benchSize = kBufferSize - (long)progressInfo.InSize;
+			long benchSize = kBufferSize - progressInfo.InSize;
 			PrintResults(dictionarySize, encodeTime, benchSize, false, 0);
 			System.out.print("     ");
 			PrintResults(dictionarySize, decodeTime, kBufferSize, true, compressedSize);
